CHANGED: Disabled support for php gettext extension, thus using php-gettext only now
Replaced all _() with __(). Developers should use __() and _ngettext() for translatable strings. The existing po files are prepared to check for these strings.
I consider this the final step for this branch. I didn't use any _ngettext(), but marked some strings where it would be appropriate with the comment //mfc (mark for change). Yet it is only cosmetics and could be done later.